backend/process-image/requirements.txt
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4